backend/requirements.txt
Flask
Flask[async]
flask_cors
langchain==0.0.325
openai==0.28.1
azure-search-documents==11.4.0b8
azure-storage-blob==12.14.1
azure-identity==1.13.0b3
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-speech==1.32.1
azure-ai-formrecognizer
pypdf==3.5.0
nltk
pandas
unstructured==0.5.11
bs4
matplotlib
PyPDF2
plotly
dash
dash_bootstrap_components
tiktoken
pymssql